fix: :bug: fix parent child stream logic
in the current version the parent child logic was not being used. This led to wrong paths in the requests because the manager customer_id was being used in the stream paths. This fixes it by propagating the actual customer ids from `CustomerHierarchyStream` and using the customer ids in the child streams. Also fixing a bug where the tap would try to get data from customer accounts that are not enabled.